refactoring(product-service) even more unit-tests (#27)
* more unit-tests

* even moremore unit-tests

* remvoed bean creation test

* fixed circular dependency

* fix

+++ product_service/src/main/java/com/onlineshop/product_service/services/ProductPictureService.java
@@ -1,8 +1,10 @@
 package com.onlineshop.product_service.services;
 
+import com.onlineshop.product_service.exceptions.ProductPictureExistsInProduct;
 import com.onlineshop.product_service.repositories.IProductPictureRepository;
 import com.onlineshop.product_service.entities.ProductPicture;
 import com.onlineshop.product_service.exceptions.ProductPictureDoesntExistException;
+import com.onlineshop.product_service.repositories.IProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,10 +16,13 @@ import java.util.List;
 public class ProductPictureService {
 
     private final IProductPictureRepository iProductPictureRepository;
+    private final IProductRepository iProductRepository;
 
     @Autowired
-    public ProductPictureService(IProductPictureRepository iProductPictureRepository) {
+    public ProductPictureService(IProductPictureRepository iProductPictureRepository,
+                                 IProductRepository iProductRepository) {
         this.iProductPictureRepository = iProductPictureRepository;
+        this.iProductRepository = iProductRepository;
     }
 
     public ProductPicture createProductPicture(ProductPicture productPicture) {
@@ -33,6 +38,9 @@ public class ProductPictureService {
     }
 
     public void deleteProductPictureById(Long id) {
+        if (this.iProductRepository.existsProductPictureInProduct(id))
+            throw new ProductPictureExistsInProduct();
+
         this.iProductPictureRepository.deleteById(id);
     }
